Update InputHandler.cs
I solved an issue with pen input. Previously, there was an issue where the input would be inconsistent and sometimes output incorrect pressure data. By tracking the delta time, I was able to discover that the delta between some events was 0 and it was these events that caused an issue. As a fix, InputHandler now tracks the delta between events and will add a 1ms delay if the delta is 0.